added fastio.hpp
